invoice_images invoice_id foreign key issue fixed

+++ database/migrations/2025_01_06_104428_create_invoice_images_table.php
@@ -10,7 +10,7 @@ class CreateInvoiceImagesTable extends Migration
     {
         Schema::create('invoice_images', function (Blueprint $table) {
             $table->id(); // Primary key
-            $table->unsignedBigInteger('invoice_id'); // Foreign key to invoices table
+            $table->unsignedInteger('invoice_id'); // Foreign key to invoices table
             $table->string('image_path'); // Path to the image
             $table->string('image_name')->nullable(); // Optional image name
             $table->timestamps();
